# openapi-v2.yaml
openapi: 3.0.3
info:
  version: 0.0.1
  title: Permissions API V2
  contact:
    name: Infratographer Authors
    url: http://github.com/infratographer
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:7603/api/v2
paths:
...
  /resources/{id}/role-bindings:
    get:
      tags:
        - role-bindings
      summary: list-role-bindings
      description: |
        list role-bindings for a resource.
        an optional query parameter `role_id` can be used to filter the results.
      operationId: listRoleBindings
      parameters:
        - name: role_id
          in: query
          schema:
            type: string
            example: permrv2-FQbFZMF74D0-WLNO-8MMb
      responses:
        "200":
          description: list-role-bindings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: permrbn-r7WU4juT0p-76JW-sJwzQ
                        role:
                          type: object
                          properties:
                            id:
                              type: string
                              example: permrv2-nDw3bVXYwHysvZDFyxh2C
                            name:
                              type: string
                              example: super_user
                        subjects:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                                example: idntusr-bailin
                              type:
                                type: string
                                example: user
                          example:
                            - id: idntusr-bailin
                              type: user
                    example:
                      - id: permrbn-r7WU4juT0p-76JW-sJwzQ
                        role:
                          id: permrv2-nDw3bVXYwHysvZDFyxh2C
                          name: super_user
                        subjects:
                          - id: idntusr-bailin
                            type: user
                      - id: permrbn-wNpgDB-ajBTyR6nB-WBcY
                        role:
                          id: permrv2-nDw3bVXYwHysvZDFyxh2C
                          name: super_user
                        subjects:
                          - id: idntusr-bailin
                            type: user
                          - id: idntusr-bailin-1
                            type: user
                          - id: idntusr-bailin-2
                            type: user
                          - id: idntusr-bailin-3
                            type: user
                          - id: idntusr-bailin-4
                            type: user
                          - id: idntusr-bailin-5
                            type: user
              examples:
                list-role-bindings:
                  value:
                    data:
                      - id: permrbn-r7WU4juT0p-76JW-sJwzQ
                        role:
                          id: permrv2-nDw3bVXYwHysvZDFyxh2C
                          name: super_user
                        subjects:
                          - id: idntusr-bailin
                            type: user
                      - id: permrbn-wNpgDB-ajBTyR6nB-WBcY
                        role:
                          id: permrv2-nDw3bVXYwHysvZDFyxh2C
                          name: super_user
                        subjects:
                          - id: idntusr-bailin
                            type: user
                          - id: idntusr-bailin-1
                            type: user
                          - id: idntusr-bailin-2
                            type: user
                          - id: idntusr-bailin-3
                            type: user
                          - id: idntusr-bailin-4
                            type: user
                          - id: idntusr-bailin-5
                            type: user
    post:
      tags:
        - role-bindings
      summary: create-role-binding
      description: |
        create a role-binding for a resource. The role-binding will grant the
        specified role to the specified subjects.
      operationId: createRoleBinding
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role_id:
                  type: string
                  example: permrv2-hoN9DD27g1tfKzjF8kj42
                subjects:
                  type: array
                  items:
                    type: object
                    properties:
                      conditions:
                        type: object
                        properties: {}
                      id:
                        type: string
                        example: idntgrp-my-subgroup
                  example:
                    - conditions: {}
                      id: idntgrp-my-subgroup
            examples:
              create-role-binding:
                value:
                  role_id: permrv2-hoN9DD27g1tfKzjF8kj42
                  subjects:
                    - conditions: {}
                      id: idntgrp-my-subgroup
      responses:
        "200":
          description: root-super-user / create-role-binding-group
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: permrbn-wNpgDB-ajBTyR6nB-WBcY
                  role:
                    type: object
                    properties:
                      id:
                        type: string
                        example: permrv2-nDw3bVXYwHysvZDFyxh2C
                      name:
                        type: string
                        example: super_user
                  subjects:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: idntusr-bailin
                        type:
                          type: string
                          example: user
                    example:
                      - id: idntusr-bailin
                        type: user
                      - id: idntusr-bailin-1
                        type: user
                      - id: idntusr-bailin-2
                        type: user
                      - id: idntusr-bailin-3
                        type: user
                      - id: idntusr-bailin-4
                        type: user
                      - id: idntusr-bailin-5
                        type: user
              examples:
                create-role-binding-group:
                  value:
                    id: permrbn-96Cy2kTlt3jtZwjEF9gaO
                    role:
                      id: permrv2-hoN9DD27g1tfKzjF8kj42
                      name: my_proj_doc_viewer
                    subjects:
                      - id: idntgrp-my-subgroup
                        type: group
                root-super-user:
                  value:
                    id: permrbn-wNpgDB-ajBTyR6nB-WBcY
                    role:
                      id: permrv2-nDw3bVXYwHysvZDFyxh2C
                      name: super_user
                    subjects:
                      - id: idntusr-bailin
                        type: user
                      - id: idntusr-bailin-1
                        type: user
                      - id: idntusr-bailin-2
                        type: user
                      - id: idntusr-bailin-3
                        type: user
                      - id: idntusr-bailin-4
                        type: user
                      - id: idntusr-bailin-5
                        type: user
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          example: tnntten-root
  /resources/{id}/role-bindings/{rb-id}:
    get:
      tags:
        - role-bindings
      summary: get-role-binding
      description: get-role-binding
      operationId: getRoleBinding
      responses:
        "200":
          description: get-role-binding
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: permrbn-pQByAAcMGUfS-ZvXPlrxp
                  role:
                    type: object
                    properties:
                      id:
                        type: string
                        example: permrv2-pUuZKw25TbCjoPnOYkA4w
                      name:
                        type: string
                        example: super-admin
                  subjects:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: idntclt-xT33dia6DUHdcDXsGRUes
                        type:
                          type: string
                          example: client
                    example:
                      - id: idntclt-xT33dia6DUHdcDXsGRUes
                        type: client
                      - id: idntgrp-root-admins
                        type: group
                      - id: idntusr-DUrM80Cg2VzLVEhpmQb7ovmW1EG1isBnmIFpLTA5N2k
                        type: user
                      - id: idntusr-FYfXJww0qhW1XrF-WKupOQCQ9Q84d-ieEamjxz1Hzhs
                        type: user
                      - id: idntusr-bailin
                        type: user
              examples:
                get-role-binding:
                  value:
                    id: permrbn-pQByAAcMGUfS-ZvXPlrxp
                    role:
                      id: permrv2-pUuZKw25TbCjoPnOYkA4w
                      name: super-admin
                    subjects:
                      - id: idntclt-xT33dia6DUHdcDXsGRUes
                        type: client
                      - id: idntgrp-root-admins
                        type: group
                      - id: idntusr-DUrM80Cg2VzLVEhpmQb7ovmW1EG1isBnmIFpLTA5N2k
                        type: user
                      - id: idntusr-FYfXJww0qhW1XrF-WKupOQCQ9Q84d-ieEamjxz1Hzhs
                        type: user
                      - id: idntusr-bailin
                        type: user
    delete:
      tags:
        - role-bindings
      summary: delete-role-binding
      description: delete-role-binding
      operationId: deleteRoleBinding
      responses:
        "200":
          description: ""
    patch:
      tags:
        - role-bindings
      summary: update-role-binding
      description: |
        update a role-binding, this will replace the subjects with the new
        subjects. note that role_id is immutable
      operationId: updateRoleBinding
      parameters:
        - name: id
          in: query
          schema:
            type: string
            example: loadbal-test-1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                subjects:
                  type: array
                  items:
                    type: object
                    properties:
                      conditions:
                        type: object
                        properties: {}
                      id:
                        type: string
                        example: idntclt-xT33dia6DUHdcDXsGRUes
                  example:
                    - id: idntclt-xT33dia6DUHdcDXsGRUes
                    - id: idntusr-DUrM80Cg2VzLVEhpmQb7ovmW1EG1isBnmIFpLTA5N2k
                    - id: idntusr-FYfXJww0qhW1XrF-WKupOQCQ9Q84d-ieEamjxz1Hzhs
                    - id: idntgrp-root-admins
            examples:
              update-role-binding:
                value:
                  subjects:
                    - id: idntclt-xT33dia6DUHdcDXsGRUes
                    - id: idntusr-DUrM80Cg2VzLVEhpmQb7ovmW1EG1isBnmIFpLTA5N2k
                    - id: idntusr-FYfXJww0qhW1XrF-WKupOQCQ9Q84d-ieEamjxz1Hzhs
                    - id: idntgrp-root-admins
      responses:
        "200":
          description: update-role-binding
          headers:
            Content-Length:
              schema:
                type: string
                example: "368"
            Date:
              schema:
                type: string
                example: Mon, 22 Apr 2024 21:52:22 GMT
            X-Request-Id:
              schema:
                type: string
                example: NzmFDigPmfVANuPeLMlMFtpykGKhGwvj
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: permrbn-pQByAAcMGUfS-ZvXPlrxp
                  role:
                    type: object
                    properties:
                      id:
                        type: string
                        example: permrv2-pUuZKw25TbCjoPnOYkA4w
                      name:
                        type: string
                        example: super-admin
                  subjects:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: idntclt-xT33dia6DUHdcDXsGRUes
                        type:
                          type: string
                          example: client
                    example:
                      - id: idntclt-xT33dia6DUHdcDXsGRUes
                        type: client
                      - id: idntusr-DUrM80Cg2VzLVEhpmQb7ovmW1EG1isBnmIFpLTA5N2k
                        type: user
                      - id: idntusr-FYfXJww0qhW1XrF-WKupOQCQ9Q84d-ieEamjxz1Hzhs
                        type: user
                      - id: idntgrp-root-admins
                        type: group
              examples:
                update-role-binding:
                  value:
                    id: permrbn-pQByAAcMGUfS-ZvXPlrxp
                    role:
                      id: permrv2-pUuZKw25TbCjoPnOYkA4w
                      name: super-admin
                    subjects:
                      - id: idntclt-xT33dia6DUHdcDXsGRUes
                        type: client
                      - id: idntusr-DUrM80Cg2VzLVEhpmQb7ovmW1EG1isBnmIFpLTA5N2k
                        type: user
                      - id: idntusr-FYfXJww0qhW1XrF-WKupOQCQ9Q84d-ieEamjxz1Hzhs
                        type: user
                      - id: idntgrp-root-admins
                        type: group
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          example: loadbal-lb1
      - name: rb-id
        in: path
        required: true
        schema:
          type: string
          example: permrbn-uRNlQEo6yh9DinKhSxL2z
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: http://localhost:8081/default/token
          scopes:
            openid: openid
            permissions-api: permissions-api
security:
  - oauth2:
      - 'openid permissions-api'
...